Bug 2021877 - Fix incorrect datatype cast in GetWindowUserTime() on 64-bit X11. r=stransky
GetWindowUserTime() was casting X11 property data to unsigned long* instead
of the correct uint32_t* (CARD32). On 64-bit systems, this caused the upper
32 bits to contain uninitialized memory, resulting in malformed startup tokens
when SynthesizeStartupToken() is used as a fallback (X11 only, when
DESKTOP_STARTUP_ID environment variable is not set).

This fix changes all related functions to use uint32_t consistently:
- GetWindowUserTime(): Return type and variable changed to uint32_t
- FindLatestUserTime(): Parameter and variable changed to uint32_t*
- SynthesizeStartupToken(): Variable changed to uint32_t, format specifier to %u

The bug only affects secondary Firefox instances on 64-bit X11 systems that
go through the remote service when DESKTOP_STARTUP_ID is not in the
environment. Wayland is unaffected as it uses XDG_ACTIVATION_TOKEN instead.

Differential Revision: https://phabricator.services.mozilla.com/D287064

+++ widget/gtk/WidgetUtilsGtk.cpp
@@ -26,6 +26,7 @@
 #include <gtk/gtk.h>
 #include <dlfcn.h>
 #include <glib.h>
+#include <inttypes.h>
 
 #ifdef MOZ_ENABLE_DBUS
 #  include "mozilla/ClearOnShutdown.h"
@@ -599,14 +600,13 @@ bool IsCancelledGError(GError* aGError) {
 }
 
 #if defined(MOZ_X11)
-static unsigned long GetWindowUserTime(GdkDisplay* aDisplay,
-                                       uintptr_t aWindow) {
+static uint32_t GetWindowUserTime(GdkDisplay* aDisplay, uintptr_t aWindow) {
   Atom actualType;
   int actualFormat;
   unsigned long numberOfItems;
   unsigned long bytesAfter;
   unsigned char* property = nullptr;
-  unsigned long userTime = 0;
+  uint32_t userTime = 0;
 
   Display* xDisplay = GDK_DISPLAY_XDISPLAY(aDisplay);
   Atom atom =
@@ -617,7 +617,7 @@ static unsigned long GetWindowUserTime(GdkDisplay* aDisplay,
                          &bytesAfter, &property) == Success &&
       property) {
     if (numberOfItems == 1) {
-      userTime = *((unsigned long*)property);
+      userTime = *((uint32_t*)property);
     }
     XFree(property);
   }
@@ -626,12 +626,12 @@ static unsigned long GetWindowUserTime(GdkDisplay* aDisplay,
 }
 
 void FindLatestUserTime(GdkDisplay* aDisplay, uintptr_t aWindow,
-                        unsigned long* aLatestTime) {
+                        uint32_t* aLatestTime) {
   Window rootReturn;
   Window parentReturn;
   Window* children;
   unsigned int numberOfChildren;
-  unsigned long userTime;
+  uint32_t userTime;
 
   Display* xDisplay = GDK_DISPLAY_XDISPLAY(aDisplay);
 
@@ -651,7 +651,7 @@ void FindLatestUserTime(GdkDisplay* aDisplay, uintptr_t aWindow,
 
 // Assume we're started from user interaction and infer user time if its missing
 nsCString SynthesizeStartupToken() {
-  unsigned long latestUserTime = 0;
+  uint32_t latestUserTime = 0;
   FindLatestUserTime(gdk_display_get_default(),
                      GDK_WINDOW_XID(gdk_get_default_root_window()),
                      &latestUserTime);
@@ -660,7 +660,7 @@ nsCString SynthesizeStartupToken() {
     return nsCString();
   }
 
-  return nsPrintfCString("%s_TIME%lu", g_get_host_name(), latestUserTime);
+  return nsPrintfCString("%s_TIME%" PRIu32, g_get_host_name(), latestUserTime);
 }
 #endif
